+++ hoomi/api/views/mobile/__init__.py
@@ -1,1 +1,1 @@
-from .user import UserCreateAPIView
+from .signup import UserCreateAPIView
