+++ hoomi/api/views/mobile/signup.py
@@ -1,4 +1,4 @@
-from django.contrib.auth import get_user_model
+from django.contrib.auth import get_user_model, authenticate
 
 from rest_framework.generics import CreateAPIView
 from rest_framework_jwt.settings import api_settings
@@ -14,3 +14,28 @@ class UserCreateAPIView(CreateAPIView):
     serializer_class = UserSerializer
     permission_classes = (AllowAny, )
     authentication_classes = ()
+
+    def post(self, request, *args, **kwargs):
+        email = request.POST.get("email")
+        password = request.POST.get("password")
+
+        user = authenticate(
+            email=email,
+            password=password,
+        )
+
+        if user:
+            response_data = {"Error": "Already Joiner"}
+
+            return Response(
+                response_data,
+                status.HTTP_401_UNAUTHORIZED,
+            )
+
+        user = self.create(request, *args, **kwargs)
+        response_data = {"Success": "Signup Success"}
+
+        return Response(
+            response_data,
+            status.HTTP_201_CREATED,
+        )
